fix: migrate lot consumption pipeline to decimal, truncate to 6dp at FlexiBee boundary
- Bump Rem.FlexiBeeSDK.Client to 0.1.135 (Amount/AmountIssued now decimal)
- Add NuGet.config with local-flexi-sdk source for local pack resolution
- IngredientRequirement.RequiredAmount: double → decimal
- ConsumptionItem.Amount: double → decimal
- FailedConsumptionItem.Amount: double → decimal
- FefoConsumptionAllocator.AllocationEpsilon: double → decimal literal (0.001m)
- FlexiIngredientRequirementAggregator: explicit (decimal) cast at BoM double→decimal boundary
- FlexiIngredientStockValidator: availableStock cast to decimal for comparison with RequiredAmount
- FlexiManufactureClient: remove now-redundant (double) cast on AdjustedConsumption assignment
- FlexiManufactureDocumentService: add RoundForFlexi(decimal) helper using MidpointRounding.ToZero
  at 6dp; apply to all Amount/AmountIssued DTO fields in consumption and production methods;
  update cost accumulation lines to (double)amount; update comments

+++ backend/src/Adapters/Anela.Heblo.Adapters.Flexi/Manufacture/Internal/FlexiIngredientRequirementAggregator.cs
@@ -26,7 +26,7 @@ internal sealed class FlexiIngredientRequirementAggregator : IFlexiIngredientReq
 
             foreach (var ingredient in template.Ingredients.Where(w => w.ProductType != ProductType.UNDEFINED))
             {
-                var scaledAmount = ingredient.Amount * scaleFactor;
+                var scaledAmount = (decimal)(ingredient.Amount * scaleFactor);
 
                 if (ingredientRequirements.TryGetValue(ingredient.ProductCode, out var existing))
                 {
